- Moved window stream logic out of main.py and into its own file window_stream.py. Now main.py is focused on just using the code, not implementing functionality. - Debug logs removed from those two files. - Tests in test_window_streams.py are rewritten with clearer comments, following the "Given/When/Then" pattern.
