+++ tests/test_window_streams.py
@@ -3,159 +3,181 @@ import numpy as np
 from neuros.main import WindowConfig, stream_windows, board_stream
 
 
-def test_window_config_validation():
+def test_window_config_creation():
     """Test window configuration validation"""
-    # Valid configurations
-    WindowConfig(window_ms=500.0, overlap_ms=250.0)  # Known good values
-    WindowConfig(window_ms=100.0, overlap_ms=60.0)  # Another known good set
+    # Given: Valid window configurations
+    valid_configs = [
+        (500.0, 250.0, "Standard window with 50% overlap"),
+        (100.0, 60.0, "Small window with partial overlap"),
+        (1000.0, 0.0, "No overlap window"),
+    ]
 
-    # Invalid configurations
-    with pytest.raises(ValueError):
-        WindowConfig(window_ms=100.0, overlap_ms=100.0)  # Equal window and overlap
-    with pytest.raises(ValueError):
-        WindowConfig(window_ms=100.0, overlap_ms=150.0)  # Overlap > window
+    # When/Then: Creating valid configurations succeeds
+    for window_ms, overlap_ms, scenario in valid_configs:
+        config = WindowConfig(window_ms=window_ms, overlap_ms=overlap_ms)
+        assert config.window_ms == window_ms, scenario
+        assert config.overlap_ms == overlap_ms, scenario
+
+    # Given: Invalid window configurations with expected error messages
+    invalid_configs = [
+        (100.0, 100.0, "Overlap must be less than window size"),
+        (100.0, 150.0, "Overlap must be less than window size"),
+        (0.0, 0.0, "Window size must be positive"),
+        (-100.0, 50.0, "Window size must be positive"),
+    ]
+
+    # When/Then: Creating invalid configurations raises ValueError with expected message
+    for window_ms, overlap_ms, expected_msg in invalid_configs:
+        with pytest.raises(ValueError, match=expected_msg):
+            WindowConfig(window_ms=window_ms, overlap_ms=overlap_ms)
 
 
-def test_basic_window_shape():
-    """Test that windows have the expected shape"""
-    config = WindowConfig(window_ms=500.0, overlap_ms=250.0)  # Known good values
+def test_sample_conversion():
+    """Test conversion from milliseconds to samples"""
+    # Given: A window configuration
+    config = WindowConfig(window_ms=500.0, overlap_ms=250.0)
+    sample_rate = 250  # Hz
+
+    # When: Converting to samples
+    window_samples, overlap_samples = config.to_samples(sample_rate)
+
+    # Then: Sample counts should match expected values
+    assert window_samples == 125, "500ms at 250Hz should be 125 samples"
+    assert overlap_samples == 62, "250ms at 250Hz should be 62 samples"
+
+
+def test_window_generation():
+    """Test basic window generation functionality"""
+    # Given: A standard window configuration
+    config = WindowConfig(window_ms=500.0, overlap_ms=250.0)
 
     with board_stream() as board:
+        # When: Generating windows
+        generator = stream_windows(board, config)
+        window = next(generator)
+
+        # Then: Window shape should match configuration
         sample_rate = board.get_sampling_rate(board_id=board.board_id)
         expected_samples = int(sample_rate * config.window_ms / 1000)
         eeg_channels = board.get_eeg_channels(board_id=board.board_id)
 
-        generator = stream_windows(board, config)
-        window = next(generator)
-
-        assert isinstance(window, np.ndarray)
-        assert window.shape == (len(eeg_channels), expected_samples)
+        assert isinstance(window, np.ndarray), "Window should be numpy array"
+        assert window.shape == (len(eeg_channels), expected_samples), "Window shape incorrect"
+        assert not np.any(np.isnan(window)), "Window should not contain NaN values"
 
 
 def test_window_overlap():
-    """Test that consecutive windows overlap by the expected amount"""
-    config = WindowConfig(window_ms=500.0, overlap_ms=250.0)  # Known good values
+    """Test window overlap behavior"""
+    # Given: A configuration with 50% overlap
+    config = WindowConfig(window_ms=500.0, overlap_ms=250.0)
 
     with board_stream() as board:
         generator = stream_windows(board, config)
 
-        # Get two consecutive windows
+        # When: Generating consecutive windows
         window1 = next(generator)
         window2 = next(generator)
 
-        # Calculate expected overlap in samples
+        # Then: Windows should overlap by the specified amount
         sample_rate = board.get_sampling_rate(board_id=board.board_id)
         overlap_samples = int(sample_rate * config.overlap_ms / 1000)
 
-        # Check that the end of window1 matches the start of window2
         np.testing.assert_array_almost_equal(
             window1[:, -overlap_samples:],
             window2[:, :overlap_samples],
-            decimal=5  # Adjust precision as needed
+            decimal=5,
+            err_msg="Overlap regions should match"
         )
 
 
-@pytest.mark.parametrize("config", [
-    WindowConfig(window_ms=500.0, overlap_ms=250.0),  # Known good
-    WindowConfig(window_ms=100.0, overlap_ms=60.0),  # Also good
-])
-def test_window_continuity(config):
-    """Test that windows contain different data and progress over time"""
+def test_data_continuity():
+    """Test that window data progresses over time"""
+    # Given: A standard configuration
+    config = WindowConfig(window_ms=500.0, overlap_ms=250.0)
+
     with board_stream() as board:
         generator = stream_windows(board, config)
 
-        # Skip first few windows to let board stabilize
+        # When: Collecting multiple windows
+        windows = []
         for _ in range(3):
-            next(generator)
+            windows.append(next(generator))
 
-        # Get test windows
-        windows = [next(generator) for _ in range(3)]
-
-        # Verify windows are different
+        # Then: Windows should contain different but valid data
         for i in range(len(windows)):
             for j in range(i + 1, len(windows)):
-                # Windows should not be identical
-                assert not np.array_equal(windows[i], windows[j])
-
-                # But they should be in a reasonable range
-                assert np.all(np.abs(windows[i]) < 1000)  # Adjust threshold as needed
-
-
-def test_eeg_channel_selection():
-    """Test that windows only contain EEG channels"""
-    config = WindowConfig(window_ms=500.0, overlap_ms=250.0)
-
-    with board_stream() as board:
-        eeg_channels = board.get_eeg_channels(board_id=board.board_id)
-        generator = stream_windows(board, config)
-        window = next(generator)
-
-        assert window.shape[0] == len(eeg_channels)
-
-
-def test_generator_cleanup():
-    """Test that the generator cleans up properly"""
-    config = WindowConfig(window_ms=500.0, overlap_ms=250.0)
-
-    with board_stream() as board:
-        generator = stream_windows(board, config)
-
-        # Get a few windows
-        for _ in range(3):
-            window = next(generator)
-            assert window is not None
+                assert not np.array_equal(windows[i], windows[j]), \
+                    "Consecutive windows should contain different data"
+                assert np.all(np.isfinite(windows[i])), \
+                    "All values should be finite"
+                assert np.all(np.abs(windows[i]) < 1000), \
+                    "Values should be within reasonable range"
 
 
 def test_error_handling():
-    """Test handling of board errors and recovery"""
+    """Test recovery from board interruption"""
+    # Given: A running board stream
     config = WindowConfig(window_ms=500.0, overlap_ms=250.0)
 
     with board_stream() as board:
         generator = stream_windows(board, config)
 
-        # Get initial window to verify normal operation
+        # Initially verify normal operation
         window1 = next(generator)
-        assert window1 is not None
-        assert isinstance(window1, np.ndarray)
+        initial_shape = window1.shape
+        assert np.all(np.isfinite(window1)), "Initial data should be valid"
 
-        # Force an error by stopping the board
+        # When: Board stream is interrupted
         board.stop_stream()
 
-        # Generator should either:
-        # 1. Return empty/zero data
-        # 2. Skip the error and continue on next valid data
-        # 3. Or maintain last valid state
-        try:
-            window2 = next(generator)
-            assert window2 is not None  # Shouldn't crash
-        except Exception as e:
-            assert str(e) != "Error during board operation: Board is not streaming"
+        # Then: Generator should handle interruption gracefully
+        window2 = next(generator)
+        assert isinstance(window2, np.ndarray), "Should return valid array type"
+        assert window2.shape == initial_shape, "Should maintain consistent shape"
+        assert np.all(np.isfinite(window2)), "Should contain valid data"
 
-        # Start the board again
+        # When: Stream is restored
         board.start_stream()
 
-        # Should recover and provide new data
+        # Then: Generator should resume normal operation
         window3 = next(generator)
-        assert window3 is not None
-        assert isinstance(window3, np.ndarray)
+        assert isinstance(window3, np.ndarray), "Should return valid array type"
+        assert window3.shape == initial_shape, "Should maintain consistent shape"
+        assert np.all(np.isfinite(window3)), "Should contain valid data"
+        assert not np.array_equal(window3, window2), "Should contain new data"
 
 
-@pytest.mark.parametrize("bad_config", [
-    WindowConfig(window_ms=10.0, overlap_ms=0.0),  # Too small
-    WindowConfig(window_ms=10000.0, overlap_ms=0.0),  # Too large
-])
-def test_extreme_window_sizes(bad_config):
-    """Test handling of extreme window sizes"""
+def test_startup_behavior():
+    """Test behavior during initial data accumulation"""
+    # Given: A window configuration requiring multiple chunks
+    config = WindowConfig(window_ms=1000.0, overlap_ms=0.0)
+
     with board_stream() as board:
-        generator = stream_windows(board, bad_config)
+        generator = stream_windows(board, config)
 
-        # Should either work or fail gracefully
-        try:
-            window = next(generator)
-            assert window is not None
-            assert isinstance(window, np.ndarray)
-        except Exception as e:
-            assert str(e) != ""  # Should have a meaningful error message
+        # When: Requesting first window
+        window = next(generator)
+
+        # Then: Should receive complete, valid window
+        sample_rate = board.get_sampling_rate(board_id=board.board_id)
+        expected_samples = int(sample_rate * config.window_ms / 1000)
+        assert window.shape[1] == expected_samples, "Should wait for complete window"
+        assert np.all(np.isfinite(window)), "Should contain valid data"
+
+
+def test_resource_cleanup():
+    """Test proper resource management"""
+    # Given: A stream configuration
+    config = WindowConfig(window_ms=500.0, overlap_ms=250.0)
+
+    # When: Using and exiting the board context
+    with board_stream() as board:
+        generator = stream_windows(board, config)
+        _ = next(generator)
+        assert board.is_prepared(), "Board should be prepared during use"
+
+    # Then: Resources should be properly cleaned up
+    assert not board.is_prepared(), "Board should be cleaned up after use"
 
 
 if __name__ == "__main__":
